HBAPP-33 Move files to forms folder
